+++ ilvyion.LaboratoryMod/DebugActions.cs
@@ -1,0 +1,29 @@
+using LudeonTK;
+using RimWorld;
+
+namespace ilvyion.LaboratoryMod;
+
+public static class DebugActions
+{
+
+    [DebugAction("ilyvion", "Hot reload languages", false, false, false, false, 0, false,
+        allowedGameStates = AllowedGameStates.Entry,
+        displayPriority = 9999)]
+    private static void HotReloadLanguages()
+    {
+        LongEventHandler.ExecuteWhenFinished(delegate
+        {
+            LanguageDatabase.Clear();
+            LanguageDatabase.InitAllMetadata();
+            GenLabel.ClearCache();
+        });
+    }
+
+    [DebugAction("ilyvion", "Hot reload languages", false, false, false, false, 0, false,
+        allowedGameStates = AllowedGameStates.Playing,
+        displayPriority = 9999)]
+    private static void HotReloadLanguagesPlaying()
+    {
+        HotReloadLanguages();
+    }
+}